# lib/api-zod/src/generated/types/createExportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Skins API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExportRequestFormat } from "./createExportRequestFormat";
import type { CreateExportRequestQuality } from "./createExportRequestQuality";

export interface CreateExportRequest {
  projectId: string;
  format: CreateExportRequestFormat;
  quality?: CreateExportRequestQuality;
}