# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Skins API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProjects: number;
  shirtProjects: number;
  pantsProjects: number;
  totalExports: number;
  aiGenerations: number;
  robloxConnected: boolean;
}