# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Skins API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
  message?: string;
}