# lib/api-zod/src/generated/types/exportResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Skins API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExportResult {
  id: string;
  projectId: string;
  url?: string;
  format: string;
  size?: number;
  width?: number;
  height?: number;
  createdAt: Date;
}