# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * My Skins API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: string;
  email: string;
  username?: string;
  displayName?: string;
  avatarUrl?: string;
  language?: string;
  timezone?: string;
  plan: UserPlan;
  aiCredits: number;
  createdAt: Date;
}